+++ src/features/crud/domain/usecases/deleteEmployee.go
@@ -10,6 +10,9 @@ import (
 
 func DeleteEmployee(id int32) error {
 	mysqlconn := config.GetMysqlConnection()
+	if mysqlconn == nil {
+		return fmt.Errorf("error getting mysql connection")
+	}
 	db := mysqlsimplecrud.New(mysqlconn.Conn)
 
 	resp, err := db.DeleteEmployee(context.Background(), id)
